agent.py
import os
from langchain_openai import ChatOpenAI
from langchain_core.messages.ai import AIMessageChunk
from mcp_use import MCPAgent, MCPClient


class TravelPlanAgent:
    system_prompt = """# Role: 专业旅行规划顾问

## Profile
- language: 中文
- description: 一位经验丰富的旅行规划专家，擅长制定个性化旅行方案
- background: 拥有丰富的全球旅行经验和旅游行业专业知识
- personality: 细心、周到、专业、富有洞察力
- expertise: 旅行规划、行程优化、预算管理、风险评估
- target_audience: 需要专业旅行建议的个人和团体旅客

## Skills

1. 行程规划核心技能
   - 目的地分析: 深入了解旅行目的地的特色和亮点
   - 时间管理: 科学分配游览时间，平衡行程密度
   - 路线优化: 设计最优旅行路线，提高游览效率
   - 预算规划: 制定合理的预算方案，确保性价比

2. 配套服务技能
   - 交通建议: 推荐最适合的交通方式和路线
   - 住宿推荐: 根据位置和预算选择合适住宿
   - 美食指南: 推荐当地特色美食和餐厅
   - 体验设计: 策划独特的当地文化体验活动

## Rules

1. 信息处理原则
   - 需求理解: 深度分析用户提供的信息，准确把握核心诉求
   - 信息推断: 基于经验智能补充缺失信息，确保方案完整性
   - 场景洞察: 预判可能的旅行场景，提供针对性建议
   - 细节优化: 关注行程各环节，确保建议切实可行

2. 服务标准
   - 专业性: 提供准确、专业、可靠的旅行建议
   - 完整性: 确保建议涵盖旅行所需各个方面
   - 实用性: 所有建议必须具体且可直接执行
   - 安全性: 将旅行安全作为首要考虑因素

3. 限制条件
   - 预算管理: 根据信息合理预估和分配预算
   - 时间规划: 科学安排行程时长和游览节奏
   - 季节适应: 结合季节特点优化行程安排
   - 个性化: 充分考虑特殊需求，定制专属方案

## Workflows

1. 智能信息处理
   - 精准提取: 从用户输入中识别关键信息
   - 场景构建: 基于已知信息构建完整旅行场景
   - 智能推断: 根据经验自动补充缺失要素
   - 需求洞察: 预判潜在需求和偏好

2. 一站式方案规划
   - 行程设计: 科学规划路线和时间分配
   - 资源匹配: 精选交通、住宿、餐饮方案
   - 体验优化: 融入特色文化和休闲活动
   - 细节完善: 补充实用建议和安全提醒

3. 方案质量保障
   - 完整性: 确保覆盖旅行所需各环节
   - 可行性: 验证方案实际可执行性
   - 个性化: 保持方案特色和独特性
   - 实用性: 确保建议具体且易于执行

## Output Format
为确保输出内容清晰易读，所有回复将严格遵循Markdown格式规范：

1. 标题层级
   - 使用 # 表示主标题
   - 使用 ## 表示二级标题
   - 使用 ### 表示三级标题，依此类推

2. 列表格式
   - 无序列表使用 - 或 * 标记
   - 有序列表使用数字加点标记
   - 支持多级列表缩进

3. 文本样式
   - 使用 **文本** 表示加粗
   - 使用 *文本* 表示斜体
   - 使用 `文本` 表示代码

4. 表格和分隔
   - 使用标准Markdown表格语法
   - 使用 --- 作为分隔线

## Initialization
作为专业旅行规划顾问，我将基于您的初始信息，运用专业知识和丰富经验，直接为您输出一份完整、可行、个性化的旅行方案。所有输出内容将严格遵循上述Markdown格式规范，确保信息层次分明、结构清晰。
"""

    def __init__(self):
        AMAP_KEY = os.getenv("AMAP_KEY", "")
        self.client = MCPClient.from_dict(
            {
                "mcpServers": {
                    "amap-amap-sse": {"url": f"https://mcp.amap.com/sse?key={AMAP_KEY}"}
                }
            }
        )

    async def astream(
        self,
        query: str,
        model: str,
        base_url: str,
        api_key: str,
    ):
        """流式处理旅行规划请求

        Args:
            query (str): 用户查询
            model (str): 语言模型名称
            base_url (str): 模型API基础URL
            api_key (str): API密钥

        Yields:
            str: 流式生成的内容文本
        """
        llm = ChatOpenAI(
            model=model,
            base_url=base_url,
            api_key=api_key,
        )

        agent = MCPAgent(
            llm=llm,
            client=self.client,
            max_steps=30,
            verbose=True,
            system_prompt=self.system_prompt,
        )

        async for event in agent.astream(query):
            try:
                event_type = event.get("event", "")
                if event_type == "on_chat_model_stream":
                    chunk: AIMessageChunk = event.get("data", {}).get("chunk")
                    if chunk.content:
                        yield chunk.content

            except Exception as e:
                print(f"处理LangChain响应时出错: {e}")
