pyproject.toml
[project]
name = "travel-plan-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-openai>=0.3.16",
    "mcp-use==1.2.12",
    "python-dotenv>=1.1.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.27.1",
    "pydantic>=2.6.3",
]
